+++ cli/CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0-alpha.12](https://github.com/tangle-network/blueprint/compare/cargo-tangle-v0.4.0-alpha.11...cargo-tangle-v0.4.0-alpha.12) - 2025-05-07
+
+### Added
+
+- *(manager)* [**breaking**] support verifying binaries with gh attestations (#938)
+
+### Other
+
+- *(ci)* fixes from attestation pr (#964)
+
+## [0.4.0-alpha.11](https://github.com/tangle-network/blueprint/compare/cargo-tangle-v0.4.0-alpha.10...cargo-tangle-v0.4.0-alpha.11) - 2025-05-06
+
+### Other
+
+- updated the following local packages: blueprint-runner, blueprint-clients, blueprint-contexts, blueprint-chain-setup, blueprint-manager, blueprint-testing-utils, blueprint-testing-utils
+
 ## [0.4.0-alpha.10](https://github.com/tangle-network/blueprint/compare/cargo-tangle-v0.4.0-alpha.9...cargo-tangle-v0.4.0-alpha.10) - 2025-05-01
 
 ### Added