+++ crates/stores/local-database/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0-alpha.3](https://github.com/tangle-network/blueprint/compare/blueprint-store-local-database-v0.1.0-alpha.2...blueprint-store-local-database-v0.1.0-alpha.3) - 2025-05-06
+
+### Added
+
+- *(pricing-engine)* use rust-decimal (#961)
+
 ## [0.1.0-alpha.2](https://github.com/tangle-network/blueprint/compare/blueprint-store-local-database-v0.1.0-alpha.1...blueprint-store-local-database-v0.1.0-alpha.2) - 2025-05-01
 
 ### Other